==> 01_Hierarchical_models.Rmd <==
---
title: "The Non-pooled model, Bayesian hierarchical models for Country-specific, Region, Income and Income-region"
author: "Oluwayomi Akinfenwa"
date: "01/10/2024"
output:
  html_document:
    toc: yes
    toc_depth: '4'
    df_print: paged
  bookdown::html_document2:
    toc: no
    toc_float: yes
    toc_depth: 4
    number_sections: yes
  pdf_document:
    toc: yes
    toc_depth: '4'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(knitr.duplicate.label = 'allow')
library(dplyr)
```


```{r , the data}
rm(list = ls()) # clear workspace


load(here::here("PISA_data", "Europe_Pisamaths.Rdata")) # Pisa_Europe_Data constructed by sourcing 00_PISA_Data.R
# sys.source("00_PISA_Data.R", envir = knitr::knit_global())
Pisa_Europe_Data_18 <- filter(Pisa_Europe_Data, year <=2018) # models fit w/o 2022

```

**We have decided to use BRMS for this analysis**

```{r}

library(dplyr)
library(brms)
library(tidybayes)
```

*Setting the BRMS model specification*

```{r}
#library(cmdstanr)
options(mc.cores = 4
    #brms.backend = "cmdstanr"
)

#Setting the seed
set.seed(1234)
bayes_seed <-1234
niter<-  12000
# niter <- 100 #use for testing purposes

```

Creating a folder to store all the models

```{r, libraries, include = FALSE}
if (!dir.exists(here::here("Manuscript_models")))
  dir.create(here::here("Manuscript_models"))
```


#### Priors for all the models

```{r}
S_Priors <- c(brms::prior(normal(500, 100), class = Intercept),
            brms::prior(normal(0, 50), class = b),
            brms::prior(cauchy(30, 100), class = sigma))

C_Priors <- c(S_Priors,
            brms::prior(cauchy(30, 10), class = sd, coef = Intercept, group=Country), 
            brms::prior(cauchy(0, 2), class = sd, coef = year_orig, group=Country))



R_Priors<- c(C_Priors,
          brms::prior(cauchy(30, 10), class = sd, coef = Intercept, group= Region),
          brms::prior(cauchy(0, 2), class = sd, coef = year_orig, group= Region)
          )

I_Priors<- c(C_Priors,
          brms::prior(cauchy(30,10), class = sd, coef = Intercept, group= Income),
          brms::prior(cauchy(0, 2), class = sd, coef = year_orig, group= Income)
          )

IR_Priors <- c(C_Priors,
          brms::prior(cauchy(30,10), class = sd, coef = Intercept, group= Income_region),
          brms::prior(cauchy(0, 2), class = sd, coef = year_orig, group= Income_region)
          )
```


#### Non-pooled model


```{r, message=FALSE}
#For the non-pooled linear model, we are interested in fitting a linear model omitting the countries with one data point.

PISA_Europe_Data_temp <- Pisa_Europe_Data_18 |> mutate(n = n(), .by=Country) |> filter(n > 1) 

model1 <- brms::brm(
  brms::bf(math ~ (year_orig*Country)),
  data = PISA_Europe_Data_temp,
  iter = niter, 
  prior = S_Priors,
  save_pars = save_pars(all = TRUE),
  control = list(adapt_delta = 0.99, max_treedepth = 25),
  silent = 2, chains = 4, seed = bayes_seed, 
  file = here::here("Manuscript_models", "model1"))


rm(PISA_Europe_Data_temp)
```


#### Level 1 hierarchical model - Country-specific model

```{r, message=FALSE}

model2 <- brms::brm(
  brms::bf(math ~ year_orig + (1 + year_orig|Country)),
  data = Pisa_Europe_Data_18,
  iter = niter, 
  prior = C_Priors,
  save_pars = save_pars(all = TRUE),
  control = list(adapt_delta = 0.93),
  chains = 4, seed = bayes_seed, 
  file = here::here("Manuscript_models", "model2"))

```


#### Level 2 hierarchical models

Region hierarchical model

```{r, message=FALSE}

model3 <- brms::brm(
  brms::bf(math ~ year_orig + (1 + year_orig|Region)+ (1 + year_orig|Country)),
  data = Pisa_Europe_Data_18,
  iter = niter, 
  prior = R_Priors,
  save_pars = save_pars(all = TRUE),
  control = list(adapt_delta = 0.99, max_treedepth = 20), silent=2,
  chains = 4, seed = bayes_seed,
  file = here::here("Manuscript_models", "model3"))

```

Income hierarchical model


```{r, message=FALSE}

model4 <- brms::brm(
  brms::bf(math ~ year_orig + (1 + year_orig|Income)+ (1 + year_orig|Country)),
  data = Pisa_Europe_Data_18,
  iter = niter, 
  prior = I_Priors,
  save_pars = save_pars(all = TRUE),
  control = list(adapt_delta = 0.997, max_treedepth = 20), silent=2,
  chains = 4, seed = bayes_seed, 
  file = here::here("Manuscript_models", "model4"))

```

Income-region hierarchical model


```{r, message=FALSE}

model5 <- brms::brm(
  brms::bf(math ~ year_orig + (1 + year_orig|Income_region)+ (1 + year_orig|Country)),
  data = Pisa_Europe_Data_18,
  iter = niter, 
  prior = IR_Priors,
  save_pars = save_pars(all = TRUE),
  control = list(adapt_delta = 0.999, max_treedepth = 15), silent=2,
  chains = 4, seed = bayes_seed,
  file = here::here("Manuscript_models", "model5"))

```

#### Model comparison tests using LOO-Compare


```{r, eval=FALSE}
fit1 <- add_criterion(model1, "loo")

fit2 <- add_criterion(model2, "loo")

fit3 <- add_criterion(model3, "loo")

fit4 <- add_criterion(model4, "loo")

fit5 <- add_criterion(model5, "loo")

#Comparing all the models
loo_compare(fit2, fit3, fit4, fit5, criterion = "loo")

rm(fit1,fit2,fit3, fit4, fit5)
```

Models are in order of 2,5,4,3 using loo_compare. But differences are small. This also is sensitive to small changes in the priors.


#### Intercepts and slopes for all 5 models as rvars

```{r}
# utility function for extracting intercept, slope from BTMS fit

extract_IS <- function(m, x,group_var, re_formula=NULL, modelID="id"){
  if (is.na(group_var))
    d <- m$data |> slice(1) 
    else d <- m$data |> group_by(.data[[group_var]]) |> slice(1) |> ungroup()
  d[[x]] <- 0
  intercepts <- tidybayes::epred_rvars(m,d, re_formula = re_formula, value="rvar") |>
    mutate(term="Intercept")
  d[[x]] <- 1
  slopes <- tidybayes::epred_rvars(m,d, re_formula = re_formula, value="rvar") |>
    mutate(term=x)
  slopes$rvar <- slopes$rvar - intercepts$rvar
  result <- bind_rows(intercepts, slopes) |> mutate(modelID=modelID)
  if (is.na(group_var))
    result |>  mutate(group="global") |> select(group, term,rvar, modelID) 
    else result |> mutate(group=.data[[group_var]]) |> select(group, term, rvar, modelID) 
}

#Country's intercept and slope estimates

countries_IS <- bind_rows(
  extract_IS(model1, "year_orig", "Country", modelID = "model1"),
  extract_IS(model2, "year_orig", "Country", modelID = "model2"),
  extract_IS(model3, "year_orig", "Country", modelID = "model3"),
  extract_IS(model4, "year_orig", "Country", modelID = "model4"),
  extract_IS(model5, "year_orig", "Country", modelID = "model5")
)
 
 

  groups_IS <- bind_rows(
   extract_IS(model2, "year_orig", group=NA, modelID = "model2", re_formula=NA),
   extract_IS(model3, "year_orig", group="Region", modelID = "model3", re_formula="~year_orig|Region") ,
   extract_IS(model4, "year_orig", "Income", modelID = "model4", re_formula="~year_orig|Income"),
   extract_IS(model5, "year_orig", "Income_region", modelID = "model5", re_formula="~year_orig|Income_region")
 )

# Country's hyper-parameter estimates, such that the country estimate is its
# respective hierarchical structure estimate.
 
groups_IS_by_country <- bind_rows(
  extract_IS(model2, "year_orig", group="Country", modelID = "model2", re_formula=NA),
  extract_IS(model3, "year_orig", group="Country", modelID = "model3", re_formula="~year_orig|Region") ,
  extract_IS(model4, "year_orig", "Country", modelID = "model4", re_formula="~year_orig|Income"),
  extract_IS(model5, "year_orig", "Country", modelID = "model5", re_formula="~year_orig|Income_region")
)

save(extract_IS, countries_IS, groups_IS, groups_IS_by_country, file=here::here("Manuscript_models", "extract_IS.Rdata"))
```
